Catching up with IWDification:
* Fixing the Wand of Monster Summoning
  https://github.com/Gibberlings3/iwdification/commit/d934e355441871c9d3b5f022fd04dcbd912cc785
* oBG2 bard kit descriptions were being wiped
  https://github.com/Gibberlings3/iwdification/commit/d6a0c94309a339444ea6afa53a412eb64fb935c9
* Apply the same targeting restrictions for arcane Energy Drain as divine:
  https://github.com/Gibberlings3/iwdification/commit/c9d744a9386b5a210cb8a80f643ba48c76148409
* Removing dupe divine spells from spellbooks
  https://github.com/Gibberlings3/iwdification/commit/26d2e239da376fc0ca1761d94f7c24c86e3ded86

On the last one, @DavidMWWallace you'll want to rewrite it into the existing spellbook update code. Basically the old versions of Physical
Mirror and Cause Serious/Critical Wounds were left in joinable NPC spellbooks. I fixed it with a sledgehammer since I couldn't muster
the enthusiasm to trace it back through the SFO calls where It Actually Does Something (tm) to fix it properly with a scalpel. Also, if
you follow that back to the G3 thread about it, the suggested function re-ordering suggested by jmerry does not fix the issue on its own.

I'm also non-committal to the Energy Drain change. I think that the arcane and divine restrictions should match, but I'm ambivalent on
whether we add the targeting restrictions to the arcane version or remove the divine ones so feel free if you want to roll these back and remove the divine restrictions.